avaliações controller

+++ src/main/java/com/example/Clubooks/books/service/LivroService.java
@@ -2,12 +2,11 @@ package com.example.Clubooks.books.service;
 
 import com.example.Clubooks.books.dto.AutoresDTO;
 import com.example.Clubooks.books.dto.BookDTO;
+import com.example.Clubooks.books.exceptions.exceptionPersonalizada;
+import com.example.Clubooks.books.model.Avaliacao;
 import com.example.Clubooks.books.model.Livro;
 import com.example.Clubooks.books.repository.LivroRepository;
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.DeserializationFeature;
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
+import com.example.Clubooks.user.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -21,6 +20,9 @@ public class LivroService {
 
     @Autowired
     private LivroRepository livroRepository;
+
+    @Autowired
+    private UserRepository userRepository;
 
     // Salva ou atualiza o livro
     public Livro salvarLivro(Livro livro) {
@@ -71,11 +73,17 @@ public class LivroService {
 
     //exclui todos os livros
     public void deletartudo() {
+        if (livroRepository.findAll().isEmpty()) {
+            throw new exceptionPersonalizada("Esta vazio ou ja foi deletado todos");
+        }
         livroRepository.deleteAll();
     }
 
     //conta todos os livros
     public long contarlivros() {
+        if (livroRepository.count() == 0) {
+            throw new exceptionPersonalizada("Esta vazio ou foi deletado todos");
+        }
         return livroRepository.count();
     }
 
@@ -112,6 +120,56 @@ public class LivroService {
 
     }
 
+    public void avaliarLivro(Double UsuarioAvaliacao, String idUsuario, String idLivro) {
+
+        var usuario = userRepository.findById(idUsuario);
+        var livroOpt = livroRepository.findById(idLivro);
+
+        if (UsuarioAvaliacao < 0 || UsuarioAvaliacao > 5 ) {
+            throw new exceptionPersonalizada("O numero não pode ser maior que 5 ou menor que 0");
+        }
+
+        if(usuario.isEmpty()) {
+            throw new exceptionPersonalizada("Usuario não encontrado " + idUsuario);
+        }
+
+        if(livroOpt.isEmpty()) {
+            throw new exceptionPersonalizada("Livro não encontrado " + idLivro);
+        }
+
+        Livro livrinho = livroOpt.get();
+
+        boolean usuarioJaAvaliou = livrinho.getAvaliacao().stream()
+                .anyMatch(avaliacao -> avaliacao.getIdUsuario().equals(idUsuario));
+
+        if (usuarioJaAvaliou) {
+            throw new exceptionPersonalizada("usuario ja avaliou");
+        }
+
+        Avaliacao novaAvaliacao = new Avaliacao();
+        novaAvaliacao.setIdUsuario(idUsuario);
+        novaAvaliacao.setNota(UsuarioAvaliacao);
+        livrinho.getAvaliacao().add(novaAvaliacao);
+
+        livroRepository.save(livrinho);
+    }
+
+    public Double mostrarAvaliacaoGeral(String idLivro){
+        var livroOpt = livroRepository.findById(idLivro);
+        if (livroOpt.isEmpty()) {
+            throw new exceptionPersonalizada("o livro não encontrado");
+        }
+        Livro livro = livroOpt.get();
+
+        if (livro.getAvaliacao().isEmpty()) {
+            throw new exceptionPersonalizada("o livro não tem nenhuma avaliação");
+        }
+
+
+        Double mediaAvaliacao = livro.getAvaliacao().stream().mapToDouble(avaliacao -> avaliacao.getNota()).average().orElse(0.0);
+        return mediaAvaliacao;
+
+    }
 
 
 }
